WIP.2 clean up
